Refactor component imports to use section-specific paths and remove unused components
