+++ src/components/landing/(sections)/HowItWorks/HowItWorksCard.tsx
