+++ src/components/landing/(sections)/Partnership/Partnership.tsx
@@ -2,7 +2,7 @@
 
 import React, { useMemo, useState } from "react";
 import { Badge } from "@/components/ui/badge";
-import PartnershipCards from "@/components/landing/PartnershipCards";
+import PartnershipCards from "@/components/landing/(sections)/Partnership/PartnershipCards";
 import { partnerships } from "@/lib/data";
 
 const categories = [
