+++ src/components/landing/(sections)/Partnership/PartnershipCards.tsx
